cargobids/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
idna==2.9
jsonfield==3.1.0
ndg-httpsclient==0.5.1
psycopg2-binary==2.8.4
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.10.5
pytz==2019.3
requests==2.23.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.0
stripe==2.43.0
swapper==1.1.2.post1
urllib3==1.25.8
